refactor: product를 store와 같은 모듈에 위치 #27
