--- api/product/src/main/java/com/backtothefuture/product/ProductApplication.java
@@ -1,16 +1,0 @@
-package com.backtothefuture.product;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
-@ComponentScan(basePackages = {"com.backtothefuture"})
-public class ProductApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ProductApplication.class, args);
-    }
-
-}
