--- api/product/src/main/resources/application.yml
@@ -1,15 +1,0 @@
-spring:
-  profiles:
-    include: domain, infra, security
-    active: dev
-  jpa:
-    open-in-view: false # view, Controller ??? Lazy? ???? ??
-    properties:
-      hibernate:
-        show_sql: true # sql query console ?? ?? ??
-        format_sql: true # console ?? ???? format ??
-
-server:
-  port: 8083 # h2 default port 가 8082
-  servlet:
-    context-path: /v1
