+++ docs/openapi/store-openapi.yaml
@@ -4,7 +4,7 @@ info:
   description: API Documentation
   version: 0.1.0
 servers:
-- url: http://localhost:8083/v1
+- url: http://localhost:8081/v1
 tags: []
 paths:
   /products/{storeId}:
@@ -23,11 +23,11 @@ paths:
           type: number
       - name: Authorization
         in: header
-        description: "JWT 인증 토큰. 'Bearer + ${Jwt Token}' 형식으로 입력해 주세요."
+        description: "JWT 인증 토큰. 'Bearer ${Jwt Token}' 형식으로 입력해 주세요."
         required: true
         schema:
           type: string
-        example: "Bearer + ${JWT Token}"
+        example: "Bearer ${JWT Token}"
       requestBody:
         content:
           application/json:
@@ -70,11 +70,11 @@ paths:
           type: number
       - name: Authorization
         in: header
-        description: "JWT 인증 토큰. 'Bearer + ${Jwt Token}' 형식으로 입력해 주세요."
+        description: "JWT 인증 토큰. 'Bearer ${Jwt Token}' 형식으로 입력해 주세요."
         required: true
         schema:
           type: string
-        example: "Bearer + ${JWT Token}"
+        example: "Bearer ${JWT Token}"
       responses:
         "204":
           description: "204"
